Move shared utils
