+++ services/rows/src/rows/routes/rows.py
@@ -6,7 +6,7 @@ import os
 import random
 import shutil
 from itertools import islice
-from typing import Any, List, Literal, Mapping, Optional, TypedDict, Union
+from typing import Any, List, Literal, Optional, Union
 
 import pyarrow as pa
 from datasets import Features
@@ -23,17 +23,19 @@ from libapi.utils import (
     are_valid_parameters,
     get_json_api_error_response,
     get_json_ok_response,
+    to_rows_list,
     try_backfill_dataset,
 )
-from libcommon.parquet_utils import Indexer, StrPath
+from libcommon.parquet_utils import Indexer
 from libcommon.processing_graph import ProcessingGraph
 from libcommon.prometheus import StepProfiler
 from libcommon.simple_cache import CachedArtifactError
+from libcommon.storage import StrPath
 from libcommon.viewer_utils.asset import (
     glob_rows_in_assets_dir,
     update_last_modified_date_of_rows_in_assets_dir,
 )
-from libcommon.viewer_utils.features import get_cell_value
+from libcommon.viewer_utils.features import to_features_list
 from starlette.requests import Request
 from starlette.responses import Response
 
@@ -43,117 +45,10 @@ logger = logging.getLogger(__name__)
 MAX_ROWS = 100
 
 
-class ParquetDataProcessingError(Exception):
-    pass
-
-
 ALL_COLUMNS_SUPPORTED_DATASETS_ALLOW_LIST: Union[Literal["all"], List[str]] = ["arabic_speech_corpus"]  # for testing
 
 # audio still has some errors when librosa is imported
 UNSUPPORTED_FEATURES_MAGIC_STRINGS = ["'binary'", "Audio("]
-
-Row = Mapping[str, Any]
-
-
-class FeatureItem(TypedDict):
-    feature_idx: int
-    name: str
-    type: Row
-
-
-# in JSON, dicts do not carry any order, so we need to return a list
-#
-# > An object is an *unordered* collection of zero or more name/value pairs, where a name is a string and a value
-#   is a string, number, boolean, null, object, or array.
-# > An array is an *ordered* sequence of zero or more values.
-# > The terms "object" and "array" come from the conventions of JavaScript.
-# from https://stackoverflow.com/a/7214312/7351594 / https://www.rfc-editor.org/rfc/rfc7159.html
-def to_features_list(features: Features) -> List[FeatureItem]:
-    features_dict = features.to_dict()
-    return [
-        {
-            "feature_idx": idx,
-            "name": name,
-            "type": features_dict[name],
-        }
-        for idx, name in enumerate(features)
-    ]
-
-
-class RowItem(TypedDict):
-    row_idx: int
-    row: Mapping[str, Any]
-    truncated_cells: List[str]
-
-
-def to_rows_list(
-    pa_table: pa.Table,
-    dataset: str,
-    config: str,
-    split: str,
-    cached_assets_base_url: str,
-    cached_assets_directory: StrPath,
-    offset: int,
-    features: Features,
-    unsupported_columns: List[str],
-) -> List[RowItem]:
-    num_rows = pa_table.num_rows
-    for idx, (column, feature) in enumerate(features.items()):
-        if column in unsupported_columns:
-            pa_table = pa_table.add_column(idx, column, pa.array([None] * num_rows))
-    # transform the rows, if needed (e.g. save the images or audio to the assets, and return their URL)
-    try:
-        transformed_rows = transform_rows(
-            dataset=dataset,
-            config=config,
-            split=split,
-            rows=pa_table.to_pylist(),
-            features=features,
-            cached_assets_base_url=cached_assets_base_url,
-            cached_assets_directory=cached_assets_directory,
-            offset=offset,
-        )
-    except Exception as err:
-        raise ParquetDataProcessingError(
-            "Server error while post-processing the split rows. Please report the issue."
-        ) from err
-    return [
-        {
-            "row_idx": idx + offset,
-            "row": row,
-            "truncated_cells": [],
-        }
-        for idx, row in enumerate(transformed_rows)
-    ]
-
-
-def transform_rows(
-    dataset: str,
-    config: str,
-    split: str,
-    rows: List[Row],
-    features: Features,
-    cached_assets_base_url: str,
-    cached_assets_directory: StrPath,
-    offset: int,
-) -> List[Row]:
-    return [
-        {
-            featureName: get_cell_value(
-                dataset=dataset,
-                config=config,
-                split=split,
-                row_idx=offset + row_idx,
-                cell=row[featureName] if featureName in row else None,
-                featureName=featureName,
-                fieldType=fieldType,
-                assets_base_url=cached_assets_base_url,
-                assets_directory=cached_assets_directory,
-            )
-            for (featureName, fieldType) in features.items()
-        }
-        for row_idx, row in enumerate(rows)
-    ]
 
 
 def _greater_or_equal(row_dir_name: str, row_idx: int, on_error: bool) -> bool:
