+++ services/worker/src/worker/job_runners/split/first_rows_from_streaming.py
@@ -30,7 +30,7 @@ from libcommon.processing_graph import ProcessingStep
 from libcommon.simple_cache import get_previous_step_or_raise
 from libcommon.storage import StrPath
 from libcommon.utils import JobInfo
-from libcommon.viewer_utils.features import get_cell_value
+from libcommon.viewer_utils.features import get_cell_value, to_features_list
 
 from worker.config import AppConfig, FirstRowsConfig
 from worker.dtos import CompleteJobResult, JobRunnerInfo, Row, SplitFirstRowsResponse
@@ -39,7 +39,6 @@ from worker.utils import (
     create_truncated_row_items,
     get_json_size,
     get_rows_or_raise,
-    to_features_list,
 )
